@@ -3,12 +3,12 @@ description: Demonstrates how to use the just_audio plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  audio_session: ^0.1.11
+  audio_session: ^0.1.14
   rxdart: ^0.27.7
   just_audio_mpv: ^0.1.6
   just_audio_windows: ^0.2.0
