@@ -52,13 +52,18 @@ void runTests() {
   }
 
   setUp(() {
-    audioSessionChannel.setMockMethodCallHandler((MethodCall methodCall) async {
+    _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+        .defaultBinaryMessenger
+        .setMockMethodCallHandler(audioSessionChannel,
+            (MethodCall methodCall) async {
       return null;
     });
   });
 
   tearDown(() {
-    audioSessionChannel.setMockMethodCallHandler(null);
+    _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+        .defaultBinaryMessenger
+        .setMockMethodCallHandler(audioSessionChannel, null);
   });
 
   test('init', () async {
@@ -346,7 +351,7 @@ void runTests() {
     stopwatch.start();
 
     var completer = Completer<dynamic>();
-    late StreamSubscription subscription;
+    late StreamSubscription<Duration> subscription;
     subscription = player.positionStream.listen((position) {
       if (position >= position1) {
         subscription.cancel();
@@ -1091,7 +1096,7 @@ void runTests() {
       playing: false,
     );
     var completer = Completer<dynamic>();
-    late StreamSubscription subscription;
+    late StreamSubscription<Duration> subscription;
     subscription = player.positionStream.listen((position) {
       expectDuration(position, Duration.zero);
       subscription.cancel();
@@ -1170,7 +1175,7 @@ void runTests() {
     );
     expect(player.currentIndex, 0);
     var completer = Completer<dynamic>();
-    late StreamSubscription subscription;
+    late StreamSubscription<Duration> subscription;
     subscription = player.positionStream.listen((position) {
       expectDuration(position, Duration.zero);
       subscription.cancel();
@@ -1743,7 +1748,7 @@ class MockWebServer {
   late HttpServer _server;
   int get port => _server.port;
 
-  Future start() async {
+  Future<void> start() async {
     _server = await HttpServer.bind(InternetAddress.loopbackIPv4, 0);
     _server.listen((request) async {
       final response = request.response;
@@ -1766,7 +1771,9 @@ class MockWebServer {
     });
   }
 
-  Future stop() => _server.close();
+  Future<void> stop() => _server.close();
 }
 
 class MyHttpOverrides extends HttpOverrides {}
+
+T? _ambiguate<T>(T? value) => value;
